restore language/ dir
