+++ python/triton/language/__init__.py
@@ -1,4 +1,4 @@
-from .base import (
+from triton.base import (
     TensorWrapper,
     bfloat16,
     block_type,
@@ -24,7 +24,11 @@ from .base import (
     is_triton_tensor,
     reinterpret,
 )
-from .core import (
+from ..core import (
+    minimum,
+    where,
+)
+from .stdlib import (
     abs,
     annotations,
     arange,
@@ -59,7 +63,6 @@ from .core import (
     max_contiguous,
     maximum,
     min,
-    minimum,
     multiple_of,
     num_programs,
     program_id,
@@ -73,7 +76,6 @@ from .core import (
     sum,
     swizzle2d,
     umulhi,
-    where,
     xor_sum,
     zeros,
     zeros_like,
@@ -96,30 +98,6 @@ from .random import (
 from . import libdevice
 
 __all__ = [
-    "TensorWrapper",
-    "bfloat16",
-    "block_type",
-    "constexpr",
-    "dtype",
-    "float16",
-    "float32",
-    "float64",
-    "float8",
-    "int1",
-    "int16",
-    "int32",
-    "int64",
-    "int8",
-    "pi32_t",
-    "pointer_type",
-    "tensor",
-    "uint16",
-    "uint32",
-    "uint64",
-    "uint8",
-    "void",
-    "is_triton_tensor",
-    "reinterpret",
     "abs",
     "annotations",
     "arange",
@@ -134,20 +112,35 @@ __all__ = [
     "atomic_xchg",
     "atomic_xor",
     "base",
+    "bfloat16",
+    "block_type",
     "broadcast",
     "broadcast_to",
     "cat",
     "cdiv",
     "clock",
     "constexpr",
+    "constexpr",
     "cos",
     "debug_barrier",
     "dequantize",
     "division",
     "dot",
+    "dtype",
     "exp",
     "fdiv",
+    "float16",
+    "float32",
+    "float64",
+    "float8",
     "globaltimer",
+    "int1",
+    "int16",
+    "int32",
+    "int64",
+    "int8",
+    "is_triton_tensor",
+    "libdevice",
     "load",
     "log",
     "max",
@@ -157,8 +150,20 @@ __all__ = [
     "minimum",
     "multiple_of",
     "num_programs",
+    "pair_uniform_to_normal",
+    "philox",
+    "philox_impl",
+    "pi32_t",
+    "pointer_type",
     "program_id",
+    "rand",
+    "rand4x",
+    "randint",
+    "randint4x",
+    "randn",
+    "randn4x",
     "ravel",
+    "reinterpret",
     "reshape",
     "sigmoid",
     "sin",
@@ -167,21 +172,18 @@ __all__ = [
     "store",
     "sum",
     "swizzle2d",
+    "tensor",
+    "TensorWrapper",
+    "triton",
+    "uint16",
+    "uint32",
+    "uint32_to_uniform_float",
+    "uint64",
+    "uint8",
     "umulhi",
+    "void",
     "where",
     "xor_sum",
     "zeros",
     "zeros_like",
-    "pair_uniform_to_normal",
-    "philox",
-    "philox_impl",
-    "rand",
-    "rand4x",
-    "randint",
-    "randint4x",
-    "randn",
-    "randn4x",
-    "triton",
-    "uint32_to_uniform_float",
-    "libdevice",
 ]
