+++ python/triton/language/random.py
