+++ python/triton/ops/blocksparse/softmax.py
@@ -1,6 +1,7 @@
 import torch
 
 import triton
+import triton.core
 import triton.language as tl
