+++ python/tutorials/04-low-memory-dropout.py
@@ -32,6 +32,7 @@ import tabulate
 import torch
 
 import triton
+import triton.core
 import triton.language as tl
