+++ python/tutorials/05-layer-norm.py
@@ -6,6 +6,7 @@ Layer Normalization
 import torch
 
 import triton
+import triton.core
 import triton.language as tl
 
 try:
