+++ python/tutorials/06-fused-attention.py
@@ -9,6 +9,7 @@ import pytest
 import torch
 
 import triton
+import triton.core
 import triton.language as tl
